Refactor ROI crop utilities and pass ROI models to backend
